Use old poi handling and print found seeds and finished runs
Runs using the static icfg took really long. As we could not observe the
output, it was unclear whether the run was stuck or whether was just
slow. As Spark was affected, we have two hypothesis: The machine itself
had problems (Johannes experiments had trouble too). Or the new POI
handling is incompatible with the observable static icfg and can cause
loops of some kind.